Basic Player Interactions FFI (#23)
* feat(ffi): get players ids
* feat(ffi): added get_player_ids
* fix: linter is silent
* feat(ffi): split packages into multiple units
