+++ squire_lib/src/lib.rs
@@ -9,7 +9,6 @@
     unreachable_pub,
 )]
 #![warn(rust_2018_idioms)]
-
 // Used in ffi
 #![cfg_attr(feature = "ffi", feature(allocator_api, slice_ptr_get))]
 
@@ -18,6 +17,11 @@
 pub mod error;
 #[cfg(feature = "ffi")]
 pub mod ffi;
+#[cfg(feature = "ffi")]
+pub mod ffi_player;
+#[cfg(feature = "ffi")]
+pub mod ffi_tournament;
+
 pub mod fluid_pairings;
 pub mod identifiers;
 pub mod operations;
