+++ squire_lib/tests/registration.rs
@@ -1,9 +1,8 @@
-
 #[cfg(test)]
 mod tests {
     use squire_lib::{
         operations::TournOp::*,
-        tournament::{Tournament, TournamentPreset}
+        tournament::{Tournament, TournamentPreset},
     };
     use uuid::Uuid;
 
@@ -15,29 +14,45 @@ mod tests {
             "Pioneer".into(),
         );
         // Reg status is respected
-        assert!(tourn.apply_op(RegisterPlayer(Uuid::new_v4().to_string())).is_ok());
+        assert!(tourn
+            .apply_op(RegisterPlayer(Uuid::new_v4().to_string()))
+            .is_ok());
         assert!(tourn.apply_op(UpdateReg(false)).is_ok());
-        assert!(tourn.apply_op(RegisterPlayer(Uuid::new_v4().to_string())).is_err());
+        assert!(tourn
+            .apply_op(RegisterPlayer(Uuid::new_v4().to_string()))
+            .is_err());
         assert!(tourn.apply_op(UpdateReg(true)).is_ok());
-        assert!(tourn.apply_op(RegisterPlayer(Uuid::new_v4().to_string())).is_ok());
+        assert!(tourn
+            .apply_op(RegisterPlayer(Uuid::new_v4().to_string()))
+            .is_ok());
         // Starting closes reg
         assert!(tourn.apply_op(Start()).is_ok());
-        assert!(tourn.apply_op(RegisterPlayer(Uuid::new_v4().to_string())).is_err());
+        assert!(tourn
+            .apply_op(RegisterPlayer(Uuid::new_v4().to_string()))
+            .is_err());
         assert!(tourn.apply_op(UpdateReg(true)).is_ok());
-        assert!(tourn.apply_op(RegisterPlayer(Uuid::new_v4().to_string())).is_ok());
+        assert!(tourn
+            .apply_op(RegisterPlayer(Uuid::new_v4().to_string()))
+            .is_ok());
         // Frozen tournament will never let people in
         assert!(tourn.apply_op(UpdateReg(true)).is_ok());
         assert!(tourn.apply_op(Freeze()).is_ok());
-        assert!(tourn.apply_op(RegisterPlayer(Uuid::new_v4().to_string())).is_err());
+        assert!(tourn
+            .apply_op(RegisterPlayer(Uuid::new_v4().to_string()))
+            .is_err());
         assert!(tourn.apply_op(UpdateReg(false)).is_err());
         assert!(tourn.apply_op(Thaw()).is_ok());
         assert!(tourn.apply_op(UpdateReg(false)).is_ok());
         assert!(tourn.apply_op(Freeze()).is_ok());
-        assert!(tourn.apply_op(RegisterPlayer(Uuid::new_v4().to_string())).is_err());
+        assert!(tourn
+            .apply_op(RegisterPlayer(Uuid::new_v4().to_string()))
+            .is_err());
         assert!(tourn.apply_op(Thaw()).is_ok());
         // Players can't join closed tournaments
         assert!(tourn.apply_op(End()).is_ok());
-        assert!(tourn.apply_op(RegisterPlayer(Uuid::new_v4().to_string())).is_err());
+        assert!(tourn
+            .apply_op(RegisterPlayer(Uuid::new_v4().to_string()))
+            .is_err());
     }
 
     #[test]
@@ -48,28 +63,44 @@ mod tests {
             "Pioneer".into(),
         );
         // Reg status is respected
-        assert!(tourn.apply_op(RegisterPlayer(Uuid::new_v4().to_string())).is_ok());
+        assert!(tourn
+            .apply_op(RegisterPlayer(Uuid::new_v4().to_string()))
+            .is_ok());
         assert!(tourn.apply_op(UpdateReg(false)).is_ok());
-        assert!(tourn.apply_op(RegisterPlayer(Uuid::new_v4().to_string())).is_err());
+        assert!(tourn
+            .apply_op(RegisterPlayer(Uuid::new_v4().to_string()))
+            .is_err());
         assert!(tourn.apply_op(UpdateReg(true)).is_ok());
-        assert!(tourn.apply_op(RegisterPlayer(Uuid::new_v4().to_string())).is_ok());
+        assert!(tourn
+            .apply_op(RegisterPlayer(Uuid::new_v4().to_string()))
+            .is_ok());
         // Starting closes reg
         assert!(tourn.apply_op(Start()).is_ok());
-        assert!(tourn.apply_op(RegisterPlayer(Uuid::new_v4().to_string())).is_err());
+        assert!(tourn
+            .apply_op(RegisterPlayer(Uuid::new_v4().to_string()))
+            .is_err());
         assert!(tourn.apply_op(UpdateReg(true)).is_ok());
-        assert!(tourn.apply_op(RegisterPlayer(Uuid::new_v4().to_string())).is_ok());
+        assert!(tourn
+            .apply_op(RegisterPlayer(Uuid::new_v4().to_string()))
+            .is_ok());
         // Frozen tournament will never let people in
         assert!(tourn.apply_op(UpdateReg(true)).is_ok());
         assert!(tourn.apply_op(Freeze()).is_ok());
-        assert!(tourn.apply_op(RegisterPlayer(Uuid::new_v4().to_string())).is_err());
+        assert!(tourn
+            .apply_op(RegisterPlayer(Uuid::new_v4().to_string()))
+            .is_err());
         assert!(tourn.apply_op(UpdateReg(false)).is_err());
         assert!(tourn.apply_op(Thaw()).is_ok());
         assert!(tourn.apply_op(UpdateReg(false)).is_ok());
         assert!(tourn.apply_op(Freeze()).is_ok());
-        assert!(tourn.apply_op(RegisterPlayer(Uuid::new_v4().to_string())).is_err());
+        assert!(tourn
+            .apply_op(RegisterPlayer(Uuid::new_v4().to_string()))
+            .is_err());
         assert!(tourn.apply_op(Thaw()).is_ok());
         // Players can't join closed tournaments
         assert!(tourn.apply_op(End()).is_ok());
-        assert!(tourn.apply_op(RegisterPlayer(Uuid::new_v4().to_string())).is_err());
+        assert!(tourn
+            .apply_op(RegisterPlayer(Uuid::new_v4().to_string()))
+            .is_err());
     }
 }
