+++ squire_sdk/src/response.rs
@@ -7,9 +7,9 @@ use rocket::{
     Response,
 };
 
+use serde::Deserialize;
 #[cfg(feature = "rocket")]
 use serde::Serialize;
-use serde::Deserialize;
 
 #[cfg(feature = "rocket")]
 use crate::error::SERIALIZER_ERROR;
